♻️ Refactor `ActivityFeedHeader` component to encapsulate all related UI elements
To streamline the components  further, `ActivityFeedHeader` now includes two sub-components: one for the organization activity feed, and another one for the user profile activity feed. Since they do not share the same logic, splitting them simplifies the code.
