+++ src/features/activity-feed/post/components/ActivityFeedHeader/index.jsx
@@ -1,93 +1,19 @@
-import { useParams, Link } from 'react-router-dom'
-import { Avatar, Box, Typography, Link as MuiLink } from '@mui/material'
-import { useEffect, useState } from 'react'
-import { api } from '../../../../../services/api'
+import { useParams } from 'react-router-dom'
+import { Box } from '@mui/material'
+import { OrganizationActivityFeedHeader } from './OrganizationActivityFeedHeader'
+import { UserProfileActivityFeedHeader } from './UserProfileActivityFeedHeader'
 
 import './style.scss'
 
 export function ActivityFeedHeader() {
     const userId = parseInt(useParams().userId, 10)
-    const [selectedMember, setSelectedMember] = useState(null)
-    const [isLoading, setIsLoading] = useState(true)
-
-    useEffect(() => {
-        const fetchUser = async () => {
-            try {
-                setIsLoading(true)
-                const res = await api(`/users/${userId}`)
-                setSelectedMember(res.data)
-            } catch {
-                console.log(`membre introuvable`)
-            } finally {
-                setIsLoading(false)
-            }
-        }
-
-        fetchUser()
-    }, [userId])
-
-    if (isLoading) {
-        return ''
-    }
-
-    if (!selectedMember) {
-        return (
-            <Box>
-                <Typography variant="body1">Utilisateur non trouvé.</Typography>
-            </Box>
-        )
-    }
 
     return (
-        <Box
-            className="c-user-badge__group"
-            sx={{
-                display: 'flex',
-                flexDirection: 'column',
-                alignItems: 'center',
-                margin: 'auto',
-                maxWidth: '300px',
-                padding: '0.5em'
-            }}
-        >
-            <MuiLink
-                component={Link}
-                to={`/${selectedMember.organization.id}/user/${selectedMember.id}`}
-            >
-                <Avatar
-                    className="c-user-badge__avatar"
-                    src={selectedMember.profilePicture}
-                    alt={selectedMember.name + selectedMember.surname}
-                    sx={{
-                        width: '100px',
-                        height: '100px',
-                        my: 1,
-                        mx: 'auto'
-                    }}
-                />
-                <Box
-                    className="c-user-badge__info"
-                    sx={{
-                        display: 'flex',
-                        flexDirection: 'column',
-                        alignItems: 'center',
-                        pb: 1
-                    }}
-                >
-                    <Typography
-                        className="c-user-badge__identity"
-                        variant="body1"
-                    >
-                        {selectedMember.name} {selectedMember.surname}
-                    </Typography>
-                    <Typography
-                        className="c-user-badge__job"
-                        variant="body1"
-                    >
-                        {selectedMember.job}
-                    </Typography>
-                </Box>
-            </MuiLink>
+        <Box className="c-feed-header" id="back-to-top-anchor">
+            {userId
+                ? <UserProfileActivityFeedHeader />
+                : <OrganizationActivityFeedHeader />
+            }
         </Box>
     )
 }
